More problems but I maybe have one final solution
Now a pointer isn't being properly handled by the C++ compiler itself so one last fallback before choosing to fail is to go to what I know. C#, it uses a table to store all references to classes and only deallocates it when none are referencing it. So I'm going to recreate it in C++ and hopefully, I'm begging, just please be the solution.